Improve documentation. FlatMapGroupsInPandas logical node to pythonLogicalOperators.scala.
